image_pipeline/chest_xray_segmentation/requirements.txt
absl-py==0.9.0
alt-model-checkpoint==1.13.0
appnope==0.1.0
astor==0.8.1
backcall==0.1.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
decorator==4.4.1
gast==0.2.2
google-api-core==1.16.0
google-auth==1.11.0
google-cloud-core==1.2.0
google-cloud-storage==1.25.0
google-pasta==0.1.8
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.26.0
h5py==2.10.0
idna==2.8
imageio==2.6.1
imgaug==0.2.9
ipython-genutils==0.2.0
ipython==7.11.1
jedi==0.16.0
keras-applications==1.0.8
keras-preprocessing==1.1.0
keras==2.3.1
kiwisolver==1.1.0
markdown==3.1.1
matplotlib==3.1.2
networkx==2.4
nibabel==2.5.1
numpy==1.18.1
opencv-python==4.1.2.30
opt-einsum==3.1.0
parso==0.6.0
pexpect==4.8.0
pickleshare==0.7.5
pillow==6.2.2
prompt-toolkit==3.0.3
protobuf==3.11.2
ptyprocess==0.6.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pydicom==1.4.1
pygments==2.5.2
pyparsing==2.4.6
pyqt5-sip==12.7.0
pyqt5==5.14.1
pynrrd==0.4.3
python-dateutil==2.8.1
pytz==2019.3
pywavelets==1.1.1
pyyaml==5.3
requests==2.22.0
rsa==4.0
scikit-image==0.16.2
scipy==1.2.1
shapely==1.7.0
six==1.14.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow==1.15.2
termcolor==1.1.0
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.8
werkzeug==0.16.1
wheel==0.34.1
wrapt==1.11.2